chore: 폴더 구조 변경에 따른 이동
