+++ src/components/page/SignUp.js
@@ -1,12 +1,26 @@
-import { useState } from "react";
-import { api } from "./api/api";
-import Input from "./Input";
+import { useEffect, useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { api } from "../../api/api";
+import Input from "../common/Input";
 
-function SignIn() {
+function SignUp() {
+  const navigate = useNavigate();
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [isEmail, setIsEmail] = useState(false);
   const [isPassword, setIsPassword] = useState(false);
+
+  const handleClick = async () => {
+    try {
+      await api.post("/auth/signup", {
+        email,
+        password,
+      });
+      navigate("/signin");
+    } catch {
+      alert("동일한 이메일 있음");
+    }
+  };
 
   const handleEmail = (e) => {
     setEmail(e.target.value);
@@ -20,33 +34,24 @@ function SignIn() {
       setIsPassword(true);
     }
   };
+
   const handleButton = () => {
     if (isEmail && isPassword) {
       return false;
     } else return true;
   };
 
-  const handleClick = async () => {
-    try {
-      const result = await api.post("/auth/signin", {
-        email,
-        password,
-      });
-      alert("로그인 성공");
-      localStorage.setItem("token", result.data.access_token);
-    } catch (error) {
-      if (error.response.status === 401) {
-        alert("이메일이 없거나 비밀번호가 일치하지 않습니다.");
-      } else if (error.response.status === 404) {
-        alert("존재하지않는 이메일입니다.");
-      }
-    }
-  };
+  useEffect(() => {
+    //값 초기화
+    setIsEmail(false);
+    setIsPassword(false);
+  }, []);
+
   return (
     <>
-      <div>
-        <h1>로그인</h1>
-      </div>
+      <header className="App-header">
+        <p>회원가입</p>
+      </header>
       <Input
         email={email}
         password={password}
@@ -54,13 +59,15 @@ function SignIn() {
         handlePw={handlePw}
       />
       <button
-        data-testid="signin-button"
+        type="submit"
+        data-testid="signup-button"
         onClick={handleClick}
         disabled={handleButton()}
       >
-        로그인
+        회원가입
       </button>
     </>
   );
 }
-export default SignIn;
+
+export default SignUp;
